feat: move speed bump and virtual traffic light
