+++ planning/behavior_velocity_speed_bump_module/plugins.xml
@@ -1,0 +1,3 @@
+<library path="behavior_velocity_speed_bump_module">
+  <class type="behavior_velocity_planner::SpeedBumpModulePlugin" base_class_type="behavior_velocity_planner::PluginInterface"/>
+</library>
