+++ planning/behavior_velocity_speed_bump_module/src/scene.cpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "scene_module/speed_bump/scene.hpp"
+#include "scene.hpp"
 
 #include "motion_utils/motion_utils.hpp"
-#include "scene_module/speed_bump/util.hpp"
 #include "tier4_autoware_utils/tier4_autoware_utils.hpp"
+#include "util.hpp"
 
 #include <rclcpp/rclcpp.hpp>
