+++ planning/behavior_velocity_speed_bump_module/src/util.cpp
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "util.hpp"
+
 #include "motion_utils/motion_utils.hpp"
 
 #include <behavior_velocity_planner_common/utilization/util.hpp>
-#include <scene_module/speed_bump/util.hpp>
 #include <tier4_autoware_utils/tier4_autoware_utils.hpp>
 
 #include <boost/geometry.hpp>
