--- planning/behavior_velocity_template_module/plugins.xml
@@ -1,3 +1,0 @@
-<library path="behavior_velocity_template_module">
-  <class type="behavior_velocity_planner::TemplateModulePlugin" base_class_type="behavior_velocity_planner::PluginInterface"/>
-</library>
