+++ planning/behavior_velocity_virtual_traffic_light_module/plugins.xml
@@ -1,0 +1,3 @@
+<library path="behavior_velocity_virtual_traffic_light_module">
+  <class type="behavior_velocity_planner::VirtualTrafficLightModulePlugin" base_class_type="behavior_velocity_planner::PluginInterface"/>
+</library>
