+++ planning/behavior_velocity_virtual_traffic_light_module/src/debug.cpp
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "scene.hpp"
+
 #include <motion_utils/motion_utils.hpp>
-#include <scene_module/virtual_traffic_light/scene.hpp>
 #include <tier4_autoware_utils/tier4_autoware_utils.hpp>
 
 using motion_utils::createStopVirtualWallMarker;
